+++ src/SmartFormat/Extensions/ConditionalFormatter.cs
@@ -12,6 +12,9 @@ using SmartFormat.Utilities;
 
 namespace SmartFormat.Extensions
 {
+    /// <summary>
+    /// A class to format primitive types with condition patterns.
+    /// </summary>
     public class ConditionalFormatter : IFormatter
     {
         private static readonly Regex _complexConditionPattern
@@ -19,8 +22,10 @@ namespace SmartFormat.Extensions
                 //   Description:      and/or    comparator     value
                 RegexOptions.IgnorePatternWhitespace | RegexOptions.Compiled);
 
+        ///<inheritdoc />
         public string[] Names { get; set; } = {"conditional", "cond", string.Empty};
 
+        ///<inheritdoc />
         public bool TryEvaluateFormat(IFormattingInfo formattingInfo)
         {
             var format = formattingInfo.Format;
@@ -39,7 +44,6 @@ namespace SmartFormat.Extensions
                 current is byte || current is short || current is int || current is long
                 || current is float || current is double || current is decimal;
             // An Enum is a number too:
-
             if (currentIsNumber == false && current != null && current.GetType().GetTypeInfo().IsEnum)
                 currentIsNumber = true;
 
@@ -149,7 +153,7 @@ namespace SmartFormat.Extensions
 
         /// <summary>
         /// Evaluates a conditional format.
-        /// Each condition must start with a comparor: "&gt;/&gt;=", "&lt;/&lt;=", "=", "!=".
+        /// Each condition must start with a comparer: "&gt;/&gt;=", "&lt;/&lt;=", "=", "!=".
         /// Conditions must be separated by either "&amp;" (AND) or "/" (OR).
         /// The conditional statement must end with a "?".
         /// Examples: